backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
aiofiles>=24.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
google-genai>=1.0.0
yt-dlp>=2024.0.0
